[CRASHES] Started Clear App Using Vulkan.hpp RAII
Crashes, probably because I didn't initiaise the Vulkan
headers features properly. Crashing line disabled and marked
with a @todo comment.

Abandoning here unless inspired as the C++ wrappers were not the
non-verbose easy mode for Vulkan experimentation I was expecting.
I had thought the trade was giving up host side memory and CPU
cycles, and adding to compile times in exchange for an easier
life, but adding a header-only library with lots of optional
features and different ways of doing things brings its own
complexities and learning curve in addition to those tradeoffs.